requirements.txt
imageio-ffmpeg==0.4.3
numpy
Pillow
pillow-lut==1.0
pyqt5-plugins==5.15.2.2.1.0
PyQt5-sip==12.8.1
PyYAML==5.4.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
scikit-image==0.17.2
scipy
sip==4.19.13
six
opencv-python
